+++ db/db_impl/db_impl.h
@@ -2284,7 +2284,9 @@ class DBImpl : public DB {
 
   // Flag to check whether Close() has been called on this DB
   bool closed_;
-
+  // save the closing status, for re-calling the close()
+  Status closing_status_;
+  // mutex for DB::Close()
   InstrumentedMutex closing_mutex_;
 
   // Conditional variable to coordinate installation of atomic flush results.